pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hybridrag"
version = "0.3.0"
description = "State-of-the-art RAG system with MongoDB Atlas and Voyage AI"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
    {name = "MongoDB", email = "devrel@mongodb.com"}
]
keywords = [
    "rag",
    "retrieval-augmented-generation",
    "mongodb",
    "atlas",
    "vector-search",
    "voyage-ai",
    "knowledge-graph",
    "llm",
    "ai",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Database",
    "Typing :: Typed",
]
dependencies = [
    # Core
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # Voyage AI
    "voyageai>=0.3.0",

    # Anthropic
    "anthropic>=0.39.0",

    # OpenAI
    "openai>=1.0.0",

    # Google Gemini
    "google-generativeai>=0.8.0",

    # MongoDB
    "pymongo>=4.6.0",
    "motor>=3.3.0",

    # Async and processing
    "tiktoken>=0.8.0",
    "tenacity>=9.0.0",
    "aioboto3>=13.2.0",
    "aiohttp>=3.11.9",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
]
ui = [
    "chainlit>=1.0.0",
    "pymupdf>=1.23.0",
]
observability = [
    "langfuse>=2.0.0",
]
evaluation = [
    "ragas>=0.2.0",
    "datasets>=2.14.0",
    "langchain-openai>=0.2.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
]
all = [
    "hybridrag[api,ui,observability,evaluation,dev]",
]

[project.urls]
Homepage = "https://github.com/romiluz13/Hybrid-Search-RAG"
Documentation = "https://github.com/romiluz13/Hybrid-Search-RAG#readme"
Repository = "https://github.com/romiluz13/Hybrid-Search-RAG"
Issues = "https://github.com/romiluz13/Hybrid-Search-RAG/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ["py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["hybridrag"]
skip = [".venv", "venv", "build", "dist"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
exclude = [
    "venv",
    ".venv",
    "build",
    "dist",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by black)
]

[tool.ruff.isort]
known-first-party = ["hybridrag"]
